added output type handling
